FIX: fix another file name error
