--- API/src/main/java/com/litcode/server/controller/codeController.java
@@ -1,28 +1,0 @@
-package com.litcode.server.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ResponseStatusException;
-
-@RestController
-@RequestMapping(controllerUtils.prefix + "/codes")
-public class codeController {
-
-  @GetMapping("/")
-  private String getCode() {
-    throw new ResponseStatusException(HttpStatus.NOT_IMPLEMENTED, "WAIT");
-  }
-
-  @PostMapping("/run")
-  private String runCode() {
-    throw new ResponseStatusException(HttpStatus.NOT_IMPLEMENTED, "WAIT");
-  }
-
-  @PostMapping("/submit")
-  private String submitCode() {
-    throw new ResponseStatusException(HttpStatus.NOT_IMPLEMENTED, "WAIT");
-  }
-}
