+++ API/src/main/java/com/litcode/server/service/CodeService.java
@@ -1,5 +1,5 @@
 package com.litcode.server.service;
 
-public class codeService {
-    
+public class CodeService {
+
 }
